+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file, according t
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+* Default schemas may now be set for custom post types at registration, using the `schemify_schema` property ([#3]).
+
+
 ## [0.1.0]
 
 * Initial public release.
@@ -11,3 +16,4 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased]: https://github.com/stevegrunwell/schemify/compare/master...develop
 [0.1.0]: https://github.com/stevegrunwell/schemify/releases/tag/v0.1.0
+[#3]: https://github.com/stevegrunwell/schemify/issues/3